fix: Bug - about us & partners not loading right
The upper parts of the pages were not loading correctly due to the use
of some animations, in about we change de whileInView for an animate to
load immediately the upper section.

And for the partners did something similar but removing the complex
animation and adding a simplier one (sad but necesary). Also changing
the whileInView for an animate

refactor: next.config & Footer

- Removed the dev option in next to stop displaying them
- Updated footer links (about us one)